+++ semana7/integracaoApi/lista-cadastro/src/Components/TelaListaPessoasCadastradas.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+
+export default class TelaListaPessoasCadastradas extends React.Component {
+    render() {
+        return (
+            <div>
+                <button onClick={this.props.irParaCadastro}> Ir para cadastro </button>
+            </div>
+        )
+    }
+}
